🚧: Refactor pdf component
